[osqp] Handle case where both G and A are one-line matrices
With scipy.sparse, trying to vstack() two one-line matrices raises the
following error:

	ValueError: blocks must be 2-D

Or with SciPy 0.13.3:

	ValueError: blocks must have rank 2

I don't think this is the desired behavior, as stacking other matrix shapes
works well. The workaround I added here is to convert A to a CSC matrix
beforehand. Then, SciPy does not complain.

I haven't checked if this behavior is still present in the latest version of
SciPy. If so, it would be worth reporting it to their issue tracker.