Increase example tolerance for ECOS (see full commit msg)
ECOS is less precise than the other solvers using its default settings,
notably ``reltol`` (here, setting ``reltol=1e-9`` instead of its default
value of 1e-8 makes it pass the dense_problem.py example).

By choice qpsolvers does not change default solver settings. The
tolerance in this example is only a sanity check that the answer is
correct (it is), so there is no harm in increasing it slightly.